Round 5 Hardening: format all crates, fix all clippy warnings (-D warnings), statistical GF2 benchmark, and non-trivial MBA crypto test
- Fix formatting across all 9 crates with cargo fmt
- Fix all Clippy warnings under -D warnings (0 warnings workspace-wide, 0 #[allow(...)])
- Upgrade test_gf2_rref_scaling_curve with 200 warmup runs + 500 timed iterations,
  reporting monotonic nanosecond scaling with stddev
- Upgrade crypto pipeline test to non-trivial Linear MBA identity ((c ^ k) + 2*(c & k)) - k
- Enforce unsafe contract on C-FFI raw pointer functions in smt-api

+++ crates/smt-parser/src/binary.rs
@@ -357,7 +357,7 @@ fn encode_op(buf: &mut Vec<u8>, op: &Op) {
         Op::IsInt => buf.push(64),
         Op::ConstArray(sort_id) => {
             buf.push(65);
-            encode_u32(buf, sort_id.0 as u32);
+            encode_u32(buf, sort_id.0);
         }
     }
 }
